// src/app/core/models/masters/jso/jso-category.ts
export interface JsoCategory {
    id: number;
    title: string;
    description: string;
    status: string;
    status_id :number;
    status_label: string;
}
export interface JsoCategoryPaginationResponse {
    current_page: number;
    total: number;
    per_page: number;
    last_page: number;
    from: number;
    data: JsoCategory[];
}

export interface JsoCategorySaveResponse {
    id: number;
    message: string;
}